move api utils to kebap case
