+++ src/lib/hooks/useBestQuote/flashMint.ts
@@ -6,8 +6,8 @@ import { FlashMintQuoteProvider } from '@indexcoop/flash-mint-sdk'
 
 import { MAINNET } from '@/constants/chains'
 import { Token } from '@/constants/tokens'
-import { getConfiguredZeroExApi } from '@/lib/utils/api/zeroExApi'
-import { getNetworkKey } from '@/lib/utils/api/zeroExUtils'
+import { getConfiguredZeroExApi } from '@/lib/utils/api/zeroex'
+import { getNetworkKey } from '@/lib/utils/api/zeroex-utils'
 import { getFullCostsInUsd, getGasCostsInUsd } from '@/lib/utils/costs'
 import { GasEstimatooor } from '@/lib/utils/gasEstimatooor'
 import { getCurrencyTokensForIndex } from '@/lib/utils/tokens'
