+++ src/lib/hooks/useTokenComponents.ts
@@ -2,7 +2,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 
 import { MAINNET } from '@/constants/chains'
 import { IndexToken, MoneyMarketIndex, Token } from '@/constants/tokens'
-import { IndexApi } from '@/lib/utils/api/indexApi'
+import { IndexApi } from '@/lib/utils/api/index-api'
 import { getAddressForToken } from '@/lib/utils/tokens'
 
 export interface SetComponent {
